refactor: modify service structure
