+++ src/main/java/com/spring/practice/rest/controller/FileController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.spring.practice.rest.service.StorageService;
+import com.spring.practice.rest.service.storage.StorageService;
 
 @RestController
 @RequestMapping("/files")
