+++ src/main/java/com/spring/practice/rest/controller/UserController.java
@@ -1,3 +1,4 @@
+
 package com.spring.practice.rest.controller;
 
 import java.util.List;
@@ -17,7 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.spring.practice.rest.domain.user.dto.UserCreate;
 import com.spring.practice.rest.domain.user.dto.UserInfo;
 import com.spring.practice.rest.domain.user.dto.UserUpdate;
-import com.spring.practice.rest.service.UserService;
+import com.spring.practice.rest.service.user.UserService;
 
 @RestController
 @RequestMapping("/users")
