+++ src/test/java/com/spring/practice/rest/service/user/impl/UserServiceImplTest.java
@@ -1,4 +1,6 @@
-package com.spring.practice.rest.service.impl;
+
+
+package com.spring.practice.rest.service.user.impl;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -20,7 +22,7 @@ import com.spring.practice.rest.domain.user.dto.UserCreate;
 import com.spring.practice.rest.domain.user.dto.UserInfo;
 import com.spring.practice.rest.domain.user.dto.UserUpdate;
 import com.spring.practice.rest.repository.user.UserRepository;
-import com.spring.practice.rest.service.UserService;
+import com.spring.practice.rest.service.user.UserService;
 
 @SpringBootTest
 @TestPropertySource("classpath:application-test.properties")
